+++ packages/terra-mixins/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added documentation on required peerDependencies
 
 1.31.0 - (June 28, 2019)
 ------------------