+++ packages/terra-status-view/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added documentation on required peerDependencies
 
 4.1.0  - (July 11, 2019)
 ------------------